requirements.txt
asgiref==3.8.1
Django==3.2
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
PyJWT==2.0.1
pytz==2021.1
setuptools==74.1.2
sqlparse==0.4.1
